+++ sample/src/main/java/testlibuiza/sample/MainActivity.java
@@ -16,7 +16,6 @@ import vn.loitp.core.base.BaseActivity;
 import vn.loitp.core.common.Constants;
 import vn.loitp.core.utilities.LActivityUtil;
 import vn.loitp.core.utilities.LLog;
-import vn.loitp.core.utilities.UizaPref;
 import vn.loitp.restapi.restclient.RestClientV2;
 import vn.loitp.restapi.restclient.RestClientV3;
 import vn.loitp.restapi.uiza.UizaServiceV2;
@@ -26,6 +25,7 @@ import vn.loitp.restapi.uiza.model.v2.auth.JsonBodyAuth;
 import vn.loitp.restapi.uiza.model.v3.UizaWorkspaceInfo;
 import vn.loitp.restapi.uiza.model.v3.authentication.gettoken.ResultGetToken;
 import vn.loitp.rxandroid.ApiSubscriber;
+import vn.loitp.uizavideov3.UizaUtil;
 import vn.loitp.uizavideov3.view.util.UizaDataV3;
 
 public class MainActivity extends BaseActivity {
@@ -57,21 +57,21 @@ public class MainActivity extends BaseActivity {
         findViewById(R.id.bt_uiza_video_cannot_slide_v2).setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
-                UizaPref.setAcitivityCanSlideIsRunning(activity, false);
+                UizaUtil.setAcitivityCanSlideIsRunning(activity, false);
                 callUizaVideoCannotSlideV2();
             }
         });
         findViewById(R.id.bt_uiza_video_slide_v2).setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
-                UizaPref.setAcitivityCanSlideIsRunning(activity, true);
+                UizaUtil.setAcitivityCanSlideIsRunning(activity, true);
                 callUizaVideoSlideV2();
             }
         });
         findViewById(R.id.bt_uiza_video_slide_v2_2).setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
-                UizaPref.setAcitivityCanSlideIsRunning(activity, true);
+                UizaUtil.setAcitivityCanSlideIsRunning(activity, true);
                 callUizaVideoSlideV2_2();
             }
         });
@@ -107,7 +107,7 @@ public class MainActivity extends BaseActivity {
         subscribe(service.auth(jsonBodyAuth), new ApiSubscriber<Auth>() {
             @Override
             public void onSuccess(Auth auth) {
-                UizaPref.setAuth(activity, auth, LSApplication.getInstance().getGson());
+                UizaUtil.setAuth(activity, auth, LSApplication.getInstance().getGson());
                 RestClientV2.addAuthorization(auth.getData().getToken());
             }
 
@@ -134,21 +134,21 @@ public class MainActivity extends BaseActivity {
     }
 
     private void callUizaVideoCannotSlideV2() {
-        UizaPref.setSlideUizaVideoEnabled(activity, false);
+        UizaUtil.setSlideUizaVideoEnabled(activity, false);
         Intent intent = new Intent(activity, V2UizaVideoIMActivity.class);
         startActivity(intent);
         LActivityUtil.tranIn(activity);
     }
 
     private void callUizaVideoSlideV2() {
-        UizaPref.setSlideUizaVideoEnabled(activity, true);
+        UizaUtil.setSlideUizaVideoEnabled(activity, true);
         Intent intent = new Intent(activity, V2UizaVideoIMActivitySlide.class);
         startActivity(intent);
         LActivityUtil.tranIn(activity);
     }
 
     private void callUizaVideoSlideV2_2() {
-        UizaPref.setSlideUizaVideoEnabled(activity, true);
+        UizaUtil.setSlideUizaVideoEnabled(activity, true);
         Intent intent = new Intent(activity, V2UizaVideoIMActivitySlide2.class);
         startActivity(intent);
         LActivityUtil.tranIn(activity);
@@ -157,7 +157,7 @@ public class MainActivity extends BaseActivity {
     //for uiza api v3
     private void authV3() {
         UizaServiceV3 service = RestClientV3.createService(UizaServiceV3.class);
-        UizaWorkspaceInfo uizaWorkspaceInfo = UizaV3Util.getUizaWorkspace(activity);
+        UizaWorkspaceInfo uizaWorkspaceInfo = UizaUtil.getUizaWorkspace(activity);
         if (uizaWorkspaceInfo == null) {
             return;
         }
