+++ uizacoresdk/src/main/java/vn/loitp/uizavideo/view/rl/video/UizaIMAVideo.java
@@ -44,7 +44,6 @@ import vn.loitp.core.utilities.LConnectivityUtil;
 import vn.loitp.core.utilities.LDialogUtil;
 import vn.loitp.core.utilities.LImageUtil;
 import vn.loitp.core.utilities.LLog;
-import vn.loitp.core.utilities.UizaPref;
 import vn.loitp.core.utilities.LScreenUtil;
 import vn.loitp.core.utilities.LSocialUtil;
 import vn.loitp.core.utilities.LUIUtil;
@@ -193,7 +192,7 @@ public class UizaIMAVideo extends RelativeLayout implements PreviewView.OnPrevie
         LUIUtil.showProgressBar(progressBar);
 
         //ko track neu play tu clicked pip
-        if (!UizaPref.getClickedPip(activity)) {
+        if (!UizaUtil.getClickedPip(activity)) {
             //cannot delete delay below, only works after 500mls
             LUIUtil.setDelay(500, new LUIUtil.DelayCallback() {
                 @Override
@@ -570,7 +569,7 @@ public class UizaIMAVideo extends RelativeLayout implements PreviewView.OnPrevie
             callback.isInitResult(true, mGetLinkPlay, mGetDetailEntity);
         }
         //ko track neu play tu clicked pip
-        if (!UizaPref.getClickedPip(activity)) {
+        if (!UizaUtil.getClickedPip(activity)) {
             //track event video_starts
             trackUiza(UizaData.getInstance().createTrackingInput(activity, Constants.EVENT_TYPE_VIDEO_STARTS));
         }
@@ -913,7 +912,7 @@ public class UizaIMAVideo extends RelativeLayout implements PreviewView.OnPrevie
         //LLog.d(TAG, "getLinkPlay");
         UizaUtil.setupRestClientV2(activity);
         UizaServiceV2 service = RestClientV2.createService(UizaServiceV2.class);
-        Auth auth = UizaPref.getAuth(activity, gson);
+        Auth auth = UizaUtil.getAuth(activity, gson);
         if (auth == null || auth.getData().getAppId() == null) {
             LDialogUtil.showDialog1Immersive(activity, activity.getString(R.string.auth_or_app_id_is_null_or_empty), new LDialogUtil.Callback1() {
                 @Override
