+++ uizacoresdk/src/main/java/vn/loitp/uizavideov3/UizaUtil.java
@@ -6,6 +6,7 @@ import android.app.Dialog;
 import android.app.KeyguardManager;
 import android.content.Context;
 import android.content.Intent;
+import android.content.SharedPreferences;
 import android.os.Build;
 import android.view.Gravity;
 import android.view.View;
@@ -29,12 +30,12 @@ import loitp.core.R;
 import vn.loitp.core.base.BaseActivity;
 import vn.loitp.core.common.Constants;
 import vn.loitp.core.utilities.LLog;
-import vn.loitp.core.utilities.UizaPref;
 import vn.loitp.core.utilities.LScreenUtil;
 import vn.loitp.restapi.restclient.RestClientTracking;
 import vn.loitp.restapi.restclient.RestClientV2;
 import vn.loitp.restapi.restclient.RestClientV3;
 import vn.loitp.restapi.uiza.UizaServiceV3;
+import vn.loitp.restapi.uiza.model.v2.auth.Auth;
 import vn.loitp.restapi.uiza.model.v2.listallentity.Subtitle;
 import vn.loitp.restapi.uiza.model.v3.UizaWorkspaceInfo;
 import vn.loitp.restapi.uiza.model.v3.authentication.gettoken.ResultGetToken;
@@ -289,17 +290,17 @@ public class UizaUtil {
 
     public static void setupRestClientV2(Activity activity) {
         if (RestClientV2.getRetrofit() == null && RestClientTracking.getRetrofit() == null) {
-            String currentApi = UizaPref.getApiEndPoint(activity);
+            String currentApi = getApiEndPoint(activity);
             if (currentApi == null || currentApi.isEmpty()) {
                 LLog.e(TAG, "setupRestClientV2 trackUiza currentApi == null || currentApi.isEmpty()");
                 return;
             }
-            String token = UizaPref.getToken(activity);
+            String token = getToken(activity);
             if (token == null || token.isEmpty()) {
                 LLog.e(TAG, "setupRestClientV2 trackUiza token==null||token.isEmpty()");
                 return;
             }
-            String currentTrackApi = UizaPref.getApiTrackEndPoint(activity);
+            String currentTrackApi = getApiTrackEndPoint(activity);
             if (currentTrackApi == null || currentTrackApi.isEmpty()) {
                 LLog.e(TAG, "setupRestClientV2 currentTrackApi == null || currentTrackApi.isEmpty()");
                 return;
@@ -350,7 +351,7 @@ public class UizaUtil {
         if (uizaWorkspaceInfo == null || uizaWorkspaceInfo.getUsername() == null || uizaWorkspaceInfo.getPassword() == null || uizaWorkspaceInfo.getUrlApi() == null) {
             throw new NullPointerException("UizaWorkspaceInfo cannot be null or empty!");
         }
-        UizaPref.setUizaWorkspaceInfo(context, uizaWorkspaceInfo);
+        setUizaWorkspaceInfo(context, uizaWorkspaceInfo);
         RestClientV3.init(Constants.PREFIXS + uizaWorkspaceInfo.getUrlApi());
     }
 
@@ -358,18 +359,18 @@ public class UizaUtil {
         if (context == null) {
             return null;
         }
-        return UizaPref.getUizaWorkspaceInfo(context);
+        return getUizaWorkspaceInfo(context);
     }
 
-    public static void setResultGetToken(Context context, ResultGetToken resultGetToken) {
-        UizaPref.setResultGetToken(context, resultGetToken);
+    /*public static void setResultGetToken(Context context, ResultGetToken resultGetToken) {
+        setResultGetToken(context, resultGetToken);
     }
 
     public static ResultGetToken getResultGetToken(Context context) {
-        return UizaPref.getResultGetToken(context);
-    }
+        return getResultGetToken(context);
+    }*/
 
-    public static String getToken(Context context) {
+    /*public static String getToken(Context context) {
         if (context == null) {
             return null;
         }
@@ -378,7 +379,7 @@ public class UizaUtil {
             return null;
         }
         return resultGetToken.getData().getToken();
-    }
+    }*/
 
     public static String getAppId(Context context) {
         if (context == null) {
@@ -447,12 +448,187 @@ public class UizaUtil {
         });
     }
 
-    /*public static void setData(Activity activity, Data data) {
-        UizaPref.setData(activity, data, new Gson());
-    }*/
-
-    public static Data getData(Activity activity) {
-        return UizaPref.getData(activity, new Gson());
-    }
     //=============================================================================END FOR UIZA V3
+
+    //=============================================================================START PREF
+    private final static String PREFERENCES_FILE_NAME = "loitp";
+    private final static String CHECK_APP_READY = "CHECK_APP_READY";
+    private final static String PRE_LOAD = "PRE_LOAD";
+    private final static String SLIDE_UIZA_VIDEO_ENABLED = "SLIDE_UIZA_VIDEO_ENABLED";
+    private final static String INDEX = "INDEX";
+    private final static String AUTH = "AUTH";
+    public final static String API_END_POINT = "API_END_POINT";
+    private final static String API_TRACK_END_POINT = "API_TRACK_END_POINT";
+    private final static String TOKEN = "TOKEN";
+    private final static String CLICKED_PIP = "CLICKED_PIP";
+    private final static String ACITIVITY_CAN_SLIDE_IS_RUNNING = "ACITIVITY_CAN_SLIDE_IS_RUNNING";
+    private final static String CLASS_NAME_OF_PLAYER = "CLASS_NAME_OF_PLAYER";
+
+    //for api v3
+    private final static String V3UIZAWORKSPACEINFO = "V3UIZAWORKSPACEINFO";
+    private final static String V3UIZATOKEN = "V3UIZATOKEN";
+    private final static String V3DATA = "V3DATA";
+    //end for api v3
+
+    //object
+    public static UizaWorkspaceInfo getUizaWorkspaceInfo(Context context) {
+        SharedPreferences pref = context.getSharedPreferences(PREFERENCES_FILE_NAME, 0);
+        return new Gson().fromJson(pref.getString(V3UIZAWORKSPACEINFO, ""), UizaWorkspaceInfo.class);
+    }
+
+    public static void setUizaWorkspaceInfo(Context context, UizaWorkspaceInfo uizaWorkspaceInfo) {
+        SharedPreferences.Editor editor = context.getSharedPreferences(PREFERENCES_FILE_NAME, 0).edit();
+        editor.putString(V3UIZAWORKSPACEINFO, new Gson().toJson(uizaWorkspaceInfo));
+        editor.apply();
+    }
+
+    public static ResultGetToken getResultGetToken(Context context) {
+        SharedPreferences pref = context.getSharedPreferences(PREFERENCES_FILE_NAME, 0);
+        return new Gson().fromJson(pref.getString(V3UIZATOKEN, ""), ResultGetToken.class);
+    }
+
+    public static void setResultGetToken(Context context, ResultGetToken resultGetToken) {
+        SharedPreferences.Editor editor = context.getSharedPreferences(PREFERENCES_FILE_NAME, 0).edit();
+        editor.putString(V3UIZATOKEN, new Gson().toJson(resultGetToken));
+        editor.apply();
+    }
+
+    /////////////////////////////////STRING
+    public static String getApiEndPoint(Context context) {
+        SharedPreferences pref = context.getSharedPreferences(PREFERENCES_FILE_NAME, 0);
+        return pref.getString(API_END_POINT, null);
+    }
+
+    public static void setApiEndPoint(Context context, String value) {
+        SharedPreferences.Editor editor = context.getSharedPreferences(PREFERENCES_FILE_NAME, 0).edit();
+        editor.putString(API_END_POINT, value);
+        editor.apply();
+    }
+
+    public static String getApiTrackEndPoint(Context context) {
+        SharedPreferences pref = context.getSharedPreferences(PREFERENCES_FILE_NAME, 0);
+        return pref.getString(API_TRACK_END_POINT, null);
+    }
+
+    public static void setApiTrackEndPoint(Context context, String value) {
+        SharedPreferences.Editor editor = context.getSharedPreferences(PREFERENCES_FILE_NAME, 0).edit();
+        editor.putString(API_TRACK_END_POINT, value);
+        editor.apply();
+    }
+
+    public static String getToken(Context context) {
+        SharedPreferences pref = context.getSharedPreferences(PREFERENCES_FILE_NAME, 0);
+        return pref.getString(TOKEN, null);
+    }
+
+    public static void setToken(Context context, String value) {
+        SharedPreferences.Editor editor = context.getSharedPreferences(PREFERENCES_FILE_NAME, 0).edit();
+        editor.putString(TOKEN, value);
+        editor.apply();
+    }
+
+    public static String getClassNameOfPlayer(Context context) {
+        SharedPreferences pref = context.getSharedPreferences(PREFERENCES_FILE_NAME, 0);
+        return pref.getString(CLASS_NAME_OF_PLAYER, null);
+    }
+
+    public static void setClassNameOfPlayer(Context context, String value) {
+        SharedPreferences.Editor editor = context.getSharedPreferences(PREFERENCES_FILE_NAME, 0).edit();
+        editor.putString(CLASS_NAME_OF_PLAYER, value);
+        editor.apply();
+    }
+    /////////////////////////////////BOOLEAN
+
+    public static Boolean getCheckAppReady(Context context) {
+        SharedPreferences prefs = context.getSharedPreferences(PREFERENCES_FILE_NAME, 0);
+        return prefs.getBoolean(CHECK_APP_READY, false);
+    }
+
+    public static void setCheckAppReady(Context context, Boolean value) {
+        SharedPreferences.Editor editor = context.getSharedPreferences(PREFERENCES_FILE_NAME, 0).edit();
+        editor.putBoolean(CHECK_APP_READY, value);
+        editor.apply();
+    }
+
+    public static Boolean getPreLoad(Context context) {
+        SharedPreferences prefs = context.getSharedPreferences(PREFERENCES_FILE_NAME, 0);
+        return prefs.getBoolean(PRE_LOAD, false);
+    }
+
+    public static void setPreLoad(Context context, Boolean value) {
+        SharedPreferences.Editor editor = context.getSharedPreferences(PREFERENCES_FILE_NAME, 0).edit();
+        editor.putBoolean(PRE_LOAD, value);
+        editor.apply();
+    }
+
+    public static Boolean getSlideUizaVideoEnabled(Context context) {
+        SharedPreferences prefs = context.getSharedPreferences(PREFERENCES_FILE_NAME, 0);
+        return prefs.getBoolean(SLIDE_UIZA_VIDEO_ENABLED, false);
+    }
+
+    public static void setSlideUizaVideoEnabled(Context context, Boolean value) {
+        SharedPreferences.Editor editor = context.getSharedPreferences(PREFERENCES_FILE_NAME, 0).edit();
+        editor.putBoolean(SLIDE_UIZA_VIDEO_ENABLED, value);
+        editor.apply();
+    }
+
+    public static Boolean getClickedPip(Context context) {
+        SharedPreferences prefs = context.getSharedPreferences(PREFERENCES_FILE_NAME, 0);
+        return prefs.getBoolean(CLICKED_PIP, false);
+    }
+
+    public static void setClickedPip(Context context, Boolean value) {
+        SharedPreferences.Editor editor = context.getSharedPreferences(PREFERENCES_FILE_NAME, 0).edit();
+        editor.putBoolean(CLICKED_PIP, value);
+        editor.apply();
+    }
+
+    public static Boolean getAcitivityCanSlideIsRunning(Context context) {
+        SharedPreferences prefs = context.getSharedPreferences(PREFERENCES_FILE_NAME, 0);
+        return prefs.getBoolean(ACITIVITY_CAN_SLIDE_IS_RUNNING, false);
+    }
+
+    public static void setAcitivityCanSlideIsRunning(Context context, Boolean value) {
+        SharedPreferences.Editor editor = context.getSharedPreferences(PREFERENCES_FILE_NAME, 0).edit();
+        editor.putBoolean(ACITIVITY_CAN_SLIDE_IS_RUNNING, value);
+        editor.apply();
+    }
+
+    /////////////////////////////////INT
+    public static int getIndex(Context context) {
+        SharedPreferences prefs = context.getSharedPreferences(PREFERENCES_FILE_NAME, 0);
+        return prefs.getInt(INDEX, Constants.NOT_FOUND);
+    }
+
+    public static void setIndex(Context context, int value) {
+        SharedPreferences.Editor editor = context.getSharedPreferences(PREFERENCES_FILE_NAME, 0).edit();
+        editor.putInt(INDEX, value);
+        editor.apply();
+    }
+
+    //Object
+    public static Auth getAuth(Context context, Gson gson) {
+        SharedPreferences pref = context.getSharedPreferences(PREFERENCES_FILE_NAME, 0);
+        String json = pref.getString(AUTH, null);
+        return gson.fromJson(json, Auth.class);
+    }
+
+    public static void setAuth(Context context, Auth auth, Gson gson) {
+        SharedPreferences.Editor editor = context.getSharedPreferences(PREFERENCES_FILE_NAME, 0).edit();
+        editor.putString(AUTH, gson.toJson(auth));
+        editor.apply();
+    }
+
+    public static Data getData(Context context, Gson gson) {
+        SharedPreferences pref = context.getSharedPreferences(PREFERENCES_FILE_NAME, 0);
+        String json = pref.getString(V3DATA, null);
+        return gson.fromJson(json, Data.class);
+    }
+
+    public static void setData(Context context, Data data, Gson gson) {
+        SharedPreferences.Editor editor = context.getSharedPreferences(PREFERENCES_FILE_NAME, 0).edit();
+        editor.putString(V3DATA, gson.toJson(data));
+        editor.apply();
+    }
+    //=============================================================================END PREF
 }
